src/ChessLib.Types/Validation/Validators/MoveValidation/CastlingRules/CastlingSquaresNotAttacked.cs
using ChessLib.Core.Types.Enums;
using ChessLib.Core.Types.Exceptions;
using ChessLib.Core.Types.Helpers;
using ChessLib.Core.Types.Interfaces;

namespace ChessLib.Core.Validation.Validators.MoveValidation.CastlingRules
{
    public class CastlingSquaresNotAttacked : IMoveRule
    {
        public MoveError Validate(in IBoard boardInfo, in ulong[][] postMoveBoard, in IMove move)
        {
            if (IsKingsPathInCheck(boardInfo.OpponentColor(), boardInfo.Occupancy, move))
            {
                return MoveError.CastleThroughCheck;
            }
            return MoveError.NoneSet;
        }
        /// <summary>
        /// Returns a boolean representing if castling path is attacked
        /// </summary>
        /// <param name="opponentColor"></param>
        /// <param name="occupancy"></param>
        /// <param name="move"></param>
        /// <returns></returns>
        protected static bool IsKingsPathInCheck(in Color opponentColor, in ulong[][] occupancy, in IMove move)
        {
            var moveToAndFromValues = move.SourceValue | move.DestinationValue;
            var squaresBetween = BoardHelpers.InBetween(move.SourceIndex, move.DestinationIndex) | moveToAndFromValues;
            while (squaresBetween != 0)
            {
                var square = BitHelpers.BitScanForward(squaresBetween);
                if (square.IsSquareAttackedByColor(opponentColor, occupancy, null)) return true;
                squaresBetween &= squaresBetween - 1;
            }
            return false;
        }
    }
}
